biometric-attendance-system/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Werkzeug
opencv-python
face-recognition
numpy
Pillow
pandas
openpyxl
psycopg2-binary
python-dotenv
